Refactor: move UiState to its own file
